soc-python/requirements.txt
flask>=3.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0
httpx>=0.27.0
python-nmap>=0.7.1
scapy>=2.5.0
